Update Huff.java
Check out the code I just added, methods writeFile and stringConvert. We needed stringConvert to get the bit Strings to be represented as integers. And I believe the writing of the characters/integers to the output file is correct. But I could also be way off on this one. And I also have no clue what the next step is.